useDirectoriesからロジックを分割
add parseTextToDirectories

update useDirectories
